requirements.txt
matplotlib~=3.9.4
pygame-ce